requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
crispy-bootstrap5==2024.2
distlib==0.3.8
Django==5.1.1
django-bootstrap4==24.4
django-crispy-forms==2.3
django-environ==0.11.2
environs==11.0.0
filelock==3.15.4
marshmallow==3.22.0
mysqlclient==2.2.7
packaging==24.1
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
python-decouple==3.8
python-dotenv==1.0.1
setuptools==74.1.0
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.3
whitenoise==6.7.0
